refactor: move uiDateText to ui utils
